#18: refactor models - replace utils in separate file
